requirements.txt
Django
requests==2.32.5
dotenv==0.9.9
pytest
pytest-django
factory_boy==3.3.3
djangorestframework
